# backend/openapi.yml
openapi: 3.1.0
info:
  version: 1.0.0
  title: Testing Nirvana API
servers:
  - url: api
components:
...
  schemas:
...
    Pagination:
      type: object
      properties:
        pages:
          type: number
        total:
          type: number
      required:
        - pages
        - total
...
    Provider:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            firstName:
              type: string
            lastName:
              type: string
            email:
              type: string
              format: email
          required:
            - id
            - firstName
            - lastName
            - email
        licenseNumber:
          type:
            - string
            - "null"
        credentials:
          type: string
        signature:
          type:
            - string
            - "null"
        serviceFeeStructure:
          type: string
          enum:
            - HOURLY
            - FLAT_RATE
            - PER_DIEM
        nssEnabled:
          type: boolean
        reviewNotes:
          type: object
          properties:
            notes:
              type: string
          required:
            - notes
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
            - PENDING
            - SUSPENDED
          default: ACTIVE
        documentIds:
          type: array
          items:
            type: string
        contractIds:
          type: array
          items:
            type: string
        contactIds:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - user
        - licenseNumber
        - credentials
        - signature
        - serviceFeeStructure
        - nssEnabled
        - reviewNotes
        - documentIds
        - contractIds
        - contactIds
        - createdAt
        - updatedAt
...
    Report:
      type: object
      properties:
        id:
          type: string
          format: uuid
        schoolId:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        therapyServiceId:
          type: string
          format: uuid
        reportType:
          type: string
          enum:
            - PROGRESS
            - ATTENDANCE
            - BILLING
            - ELIGIBILITY
        content:
          type: string
        createdAt:
          type: string
          default: 2025-03-27T21:45:09.629Z
        updatedAt:
          type: string
        deletedAt:
          type:
            - string
            - "null"
        school:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
          required:
            - id
            - name
        student:
          type: object
          properties:
            id:
              type: string
              format: uuid
            firstName:
              type: string
            lastName:
              type: string
          required:
            - id
            - firstName
            - lastName
        therapyService:
          type: object
          properties:
            id:
              type: string
              format: uuid
            serviceType:
              type: string
              enum:
                - SPEECH
                - OCCUPATIONAL
                - PHYSICAL
          required:
            - id
            - serviceType
      required:
        - id
        - schoolId
        - studentId
        - therapyServiceId
        - reportType
        - content
        - updatedAt
        - deletedAt
...
paths:
...
  /providers:
    get:
      description: Get a list of providers
      parameters:
        - schema:
            type: string
          required: false
          name: name
          in: query
        - schema:
            type: string
          required: false
          name: licenseNumber
          in: query
        - schema:
            type: string
          required: false
          name: credentials
          in: query
        - schema:
            type: boolean
          required: false
          name: nssEnabled
          in: query
        - schema:
            type: string
            enum:
              - ACTIVE
              - INACTIVE
              - PENDING
              - SUSPENDED
          required: false
          name: status
          in: query
        - schema:
            type: string
            enum:
              - HOURLY
              - FLAT_RATE
              - PER_DIEM
          required: false
          name: serviceFeeStructure
          in: query
        - schema:
            type: string
            enum:
              - name
              - licenseNumber
              - credentials
              - nssEnabled
              - serviceFeeStructure
              - status
            default: name
          required: false
          name: sortBy
          in: query
        - schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
          required: false
          name: sortOrder
          in: query
        - schema:
            type: number
            minimum: 1
            default: 20
          required: false
          name: perPage
          in: query
        - schema:
            type: number
            minimum: 1
            default: 1
          required: false
          name: page
          in: query
      responses:
        "200":
          description: Object with data and pagination metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Provider"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
                required:
                  - data
                  - pagination
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
    post:
      description: Create a provider
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                licenseNumber:
                  type: string
                credentials:
                  type: string
                signature:
                  type:
                    - string
                    - "null"
                serviceFeeStructure:
                  type: string
                  enum:
                    - HOURLY
                    - FLAT_RATE
                    - PER_DIEM
                  default: HOURLY
                nssEnabled:
                  type: boolean
                reviewNotes:
                  type: object
                  properties:
                    notes:
                      type: string
                  required:
                    - notes
                status:
                  type: string
                  enum:
                    - ACTIVE
                    - INACTIVE
                    - PENDING
                    - SUSPENDED
                  default: ACTIVE
              required:
                - userId
                - licenseNumber
                - credentials
                - signature
                - nssEnabled
                - reviewNotes
      responses:
        "204":
          description: The provider has been created successfully.
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
  /providers/{id}:
    get:
      description: Get a provider
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Object with data containing provider information.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Provider"
                required:
                  - data
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
    patch:
      description: Update a provider
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                licenseNumber:
                  type: string
                credentials:
                  type: string
                signature:
                  type:
                    - string
                    - "null"
                serviceFeeStructure:
                  type: string
                  enum:
                    - HOURLY
                    - FLAT_RATE
                    - PER_DIEM
                  default: HOURLY
                nssEnabled:
                  type: boolean
                reviewNotes:
                  type: object
                  properties:
                    notes:
                      type: string
                  required:
                    - notes
                status:
                  type: string
                  enum:
                    - ACTIVE
                    - INACTIVE
                    - PENDING
                    - SUSPENDED
                  default: ACTIVE
      responses:
        "204":
          description: The provider has been updated successfully.
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
    delete:
      description: Delete a provider
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "204":
          description: The provider has been deleted successfully.
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
...
  /reports:
    get:
      description: Get a list of reports
      parameters:
        - schema:
            type: string
            enum:
              - PROGRESS
              - ATTENDANCE
              - BILLING
              - ELIGIBILITY
          required: false
          name: reportType
          in: query
        - schema:
            type: string
            format: uuid
          required: false
          name: therapyServiceId
          in: query
        - schema:
            type: string
          required: false
          name: schoolName
          in: query
        - schema:
            type: string
          required: false
          name: studentName
          in: query
        - schema:
            type: string
            enum:
              - SPEECH
              - OCCUPATIONAL
              - PHYSICAL
          required: false
          name: therapyServiceType
          in: query
        - schema:
            type:
              - string
              - "null"
          required: false
          name: createdAt
          in: query
        - schema:
            type: string
            enum:
              - reportType
              - createdAt
              - updatedAt
              - content
              - schoolName
              - studentName
              - therapyServiceType
            default: createdAt
          required: false
          name: sortBy
          in: query
        - schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
          required: false
          name: sortOrder
          in: query
        - schema:
            type: number
            minimum: 1
            default: 20
          required: false
          name: perPage
          in: query
        - schema:
            type: number
            minimum: 1
            default: 1
          required: false
          name: page
          in: query
      responses:
        "200":
          description: Object with data and pagination metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Report"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
                required:
                  - data
                  - pagination
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
    post:
      description: Create a report
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                schoolId:
                  type: string
                  format: uuid
                studentId:
                  type: string
                  format: uuid
                therapyServiceId:
                  type: string
                  format: uuid
                reportType:
                  type: string
                  enum:
                    - PROGRESS
                    - ATTENDANCE
                    - BILLING
                    - ELIGIBILITY
                content:
                  type: string
                createdAt:
                  type: string
                  default: &a2 2025-03-27T21:45:09.629Z
              required:
                - schoolId
                - studentId
                - therapyServiceId
                - reportType
                - content
      responses:
        "201":
          description: The report has been created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                required:
                  - id
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
  /reports/{id}:
    get:
      description: Get a report
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Object with data containing report information.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Report"
                required:
                  - data
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
    patch:
      description: Update a report
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                schoolId:
                  type: string
                  format: uuid
                studentId:
                  type: string
                  format: uuid
                therapyServiceId:
                  type: string
                  format: uuid
                reportType:
                  type: string
                  enum:
                    - PROGRESS
                    - ATTENDANCE
                    - BILLING
                    - ELIGIBILITY
                content:
                  type: string
                createdAt:
                  type: string
                  default: *a2
      responses:
        "204":
          description: The report has been updated successfully.
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
    delete:
      description: Delete a report
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "204":
          description: The report has been deleted successfully.
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
...